Created new script for model ageing work Created io module for scripts to make loading (and eventually saving) simpler Made scripts directory a module for importing. Shutdown bash script runs as sudo.

+++ shutdown_when_complete.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ $EUID != 0 ]; then
+    sudo "$0" "$@"
+    exit $?
+fi
+
 if [ -f "/home/gabriel/signal/shut_me_down_goddamnit" ]
     then
     rm /home/gabriel/signal/shut_me_down_goddamnit
